+++ src/test/cli.test.ts
@@ -6,6 +6,8 @@
 import * as assert from 'assert';
 import * as path from 'path';
 import * as cp from 'child_process';
+import * as fs from 'fs';
+import * as os from 'os';
 
 const pkg = require('../../package.json');
 
@@ -14,19 +16,26 @@ const buildKitOptions = [
 	{ text: 'BuildKit', options: { useBuildKit: true }, },
 ] as const;
 
+interface UpResult {
+	outcome: string;
+	containerId: string;
+	composeProjectName: string | undefined;
+}
+
 describe('Dev Containers CLI', function () {
 	this.timeout('120s');
 
 	const tmp = path.relative(process.cwd(), path.join(__dirname, 'tmp'));
 	const cli = `npx --prefix ${tmp} devcontainer`;
-	async function devContainerUp(workspaceFolder: string, options?: { useBuildKit?: boolean }) {
+	async function devContainerUp(workspaceFolder: string, options?: { useBuildKit?: boolean, userDataFolder?: string }): Promise<UpResult> {
 		const buildkitOption = (options?.useBuildKit ?? false) ? '' : ' --buildkit=never';
-		const res = await shellExec(`${cli} up --workspace-folder ${workspaceFolder}${buildkitOption}`);
+		const userDataFolderOption = (options?.userDataFolder ?? false) ? ` --user-data-folder=${options?.userDataFolder}` : '';
+		const res = await shellExec(`${cli} up --workspace-folder ${workspaceFolder}${buildkitOption}${userDataFolderOption}`);
 		const response = JSON.parse(res.stdout);
 		assert.equal(response.outcome, 'success');
-		const { containerId, composeProjectName } = response;
+		const { outcome, containerId, composeProjectName } = response as UpResult;
 		assert.ok(containerId, 'Container id not found.');
-		return { containerId, composeProjectName };
+		return { outcome, containerId, composeProjectName };
 	}
 	async function devContainerDown(options: { containerId?: string | null; composeProjectName?: string | null }) {
 		if (options.containerId) {
@@ -36,6 +45,32 @@ describe('Dev Containers CLI', function () {
 			await shellExec(`docker compose --project-name ${options.composeProjectName} down`);
 		}
 	}
+	async function devContainerStop(options: { containerId?: string | null; composeProjectName?: string | null }) {
+		if (options.containerId) {
+			await shellExec(`docker stop ${options.containerId}`);
+		}
+		if (options.composeProjectName) {
+			await shellExec(`docker compose --project-name ${options.composeProjectName} stop`);
+		}
+	}
+	async function pathExists(workspaceFolder: string, location: string) {
+		try {
+			await shellExec(`${cli} exec --workspace-folder ${workspaceFolder} test -e ${location}`);
+			return true;
+		} catch (err) {
+			return false;
+		}
+	}
+	async function commandMarkerTests(workspaceFolder: string, expected: { postCreate: boolean; postStart: boolean; postAttach: boolean }, message: string) {
+		const actual = {
+			postCreate: await pathExists(workspaceFolder, '/tmp/postCreateCommand.testmarker'),
+			postStart: await pathExists(workspaceFolder, '/tmp/postStartCommand.testmarker'),
+			postAttach: await pathExists(workspaceFolder, '/tmp/postAttachCommand.testmarker'),
+		};
+		assert.deepStrictEqual(actual, expected, message);
+	}
+
+
 	before('Install', async () => {
 		await shellExec(`rm -rf ${tmp}/node_modules`);
 		await shellExec(`mkdir -p ${tmp}`);
@@ -48,11 +83,42 @@ describe('Dev Containers CLI', function () {
 	});
 
 	describe('Command build', () => {
-		it('should execute successfully with valid config', async () => {
-			const testFolder = `${__dirname}/configs/image`;
-			const res = await shellExec(`${cli} build --workspace-folder ${testFolder}`);
-			const response = JSON.parse(res.stdout);
-			assert.equal(response.outcome, 'success');
+
+		buildKitOptions.forEach(({ text, options }) => {
+			it(`should execute successfully with valid image config  [${text}]`, async () => {
+				const testFolder = `${__dirname}/configs/image`;
+				const buildKitOption = (options?.useBuildKit ?? false) ? '' : ' --buildkit=never';
+				const res = await shellExec(`${cli} build --workspace-folder ${testFolder}${buildKitOption}`);
+				const response = JSON.parse(res.stdout);
+				assert.equal(response.outcome, 'success');
+			});
+			it(`should execute successfully with valid docker-compose (image) config [${text}]`, async () => {
+				const testFolder = `${__dirname}/configs/compose-image-with-features`;
+				const buildKitOption = (options?.useBuildKit ?? false) ? '' : ' --buildkit=never';
+				const res = await shellExec(`${cli} build --workspace-folder ${testFolder}${buildKitOption}`);
+				console.log(res.stdout);
+				const response = JSON.parse(res.stdout);
+				assert.equal(response.outcome, 'success');
+			});
+			it(`should execute successfully with valid image config and extra cacheFrom [${text}]`, async () => {
+				const testFolder = `${__dirname}/configs/image`;
+				const buildKitOption = (options?.useBuildKit ?? false) ? '' : ' --buildkit=never';
+				// validate the build succeeds with an extra cacheFrom that isn't found
+				// (example scenario: CI builds for PRs)
+				const res = await shellExec(`${cli} build --workspace-folder ${testFolder} --cache-from ghcr.io/devcontainers/notFound ${buildKitOption}`);
+				const response = JSON.parse(res.stdout);
+				assert.equal(response.outcome, 'success');
+			});
+			it(`should execute successfully with valid docker-compose (image) config and extra cacheFrom [${text}]`, async () => {
+				const testFolder = `${__dirname}/configs/compose-image-with-features`;
+				const buildKitOption = (options?.useBuildKit ?? false) ? '' : ' --buildkit=never';
+				// validate the build succeeds with an extra cacheFrom that isn't found
+				// (example scenario: CI builds for PRs)
+				const res = await shellExec(`${cli} build --workspace-folder ${testFolder} --cache-from ghcr.io/devcontainers/notFound ${buildKitOption}`);
+				console.log(res.stdout);
+				const response = JSON.parse(res.stdout);
+				assert.equal(response.outcome, 'success');
+			});
 		});
 
 		it('should fail with "not found" error when config is not found', async () => {
@@ -68,6 +134,8 @@ describe('Dev Containers CLI', function () {
 			}
 			assert.equal(success, false, 'expect non-successful call');
 		});
+
+
 	});
 
 	describe('Command up', () => {
@@ -102,6 +170,105 @@ describe('Dev Containers CLI', function () {
 			}
 			assert.equal(success, false, 'expect non-successful call');
 		});
+
+		// docker-compose variations _without_ features are here (under 'up' tests)
+		// docker-compose variations _with_ features are under 'exec' to test features are installed
+		describe('for docker-compose with image without features', () => {
+			let upResult: UpResult | null = null;
+			const testFolder = `${__dirname}/configs/compose-image-without-features`;
+			before(async () => {
+				// build and start the container
+				upResult = await devContainerUp(testFolder);
+			});
+			after(async () => await devContainerDown({ composeProjectName: upResult?.composeProjectName }));
+			it('should succeed', () => {
+				assert.equal(upResult!.outcome, 'success');
+			});
+		});
+		describe('for docker-compose with Dockerfile without features', () => {
+			let upResult: UpResult | null = null;
+			const testFolder = `${__dirname}/configs/compose-Dockerfile-without-features`;
+			before(async () => {
+				// build and start the container
+				upResult = await devContainerUp(testFolder);
+			});
+			after(async () => await devContainerDown({ composeProjectName: upResult?.composeProjectName }));
+			it('should succeed', () => {
+				assert.equal(upResult!.outcome, 'success');
+			});
+		});
+
+		// Additional tests to verify the handling of persisted files
+		describe('for docker-compose with Dockerfile with features', () => {
+			describe('with existing container and persisted override files', () => {
+				let upResult1: UpResult | null = null;
+				let upResult2: UpResult | null = null;
+				let userDataFolder: string | null = null;
+				const testFolder = `${__dirname}/configs/compose-Dockerfile-with-features`;
+				before(async () => {
+					// Create a new temp folder for persisted files for this test
+					// so that we can check the contents...
+					const tmpDir = os.tmpdir();
+					userDataFolder = fs.mkdtempSync(path.join(tmpDir, 'dc-cli-test-'));
+
+					// build and start the container
+					upResult1 = await devContainerUp(testFolder, { userDataFolder });
+
+					// stop the container but don't delete it
+					await shellExec(`docker compose --project-name ${upResult1.composeProjectName} stop`);
+
+					// restart the container
+					upResult2 = await devContainerUp(testFolder, { userDataFolder });
+
+				});
+				after(async () => await devContainerDown({ composeProjectName: upResult2?.composeProjectName }));
+				it('should succeed', () => {
+					assert.equal(upResult2?.outcome, 'success');
+				});
+				it('should re-used stopped container', () => {
+					assert.equal(upResult2?.containerId, upResult1?.containerId);
+				});
+				it('should re-used the persisted override file', async () => {
+					const userDataFiles = fs.readdirSync(path.join(userDataFolder!, 'docker-compose'));
+					assert.equal(userDataFiles.length, 2); // build override and start override
+					assert.ok(userDataFiles.findIndex(f => f.startsWith('docker-compose.devcontainer.build-')) >= 0);
+					assert.ok(userDataFiles.findIndex(f => f.startsWith('docker-compose.devcontainer.containerFeatures-')) >= 0);
+				});
+			});
+			describe('with existing container and without persisted override files', () => {
+				let upResult1: UpResult | null = null;
+				let upResult2: UpResult | null = null;
+				const testFolder = `${__dirname}/configs/compose-Dockerfile-with-features`;
+				before(async () => {
+					// Create a new temp folder for persisted files for this test
+					// so that we can delete them and check all works ok
+					const tmpDir = os.tmpdir();
+					const userDataFolder = fs.mkdtempSync(path.join(tmpDir, 'dc-cli-test-'));
+
+					// build and start the container
+					upResult1 = await devContainerUp(testFolder, { userDataFolder });
+
+					// stop the container but don't delete it
+					await shellExec(`docker compose --project-name ${upResult1.composeProjectName} stop`);
+					assert.ok(upResult1?.composeProjectName);
+
+					// recreate directory to delete cached files
+					fs.rmSync(userDataFolder, { force: true, recursive: true });
+					fs.mkdirSync(userDataFolder);
+
+					// restart the container
+					upResult2 = await devContainerUp(testFolder, { userDataFolder });
+
+				});
+				after(async () => await devContainerDown({ composeProjectName: upResult2?.composeProjectName }));
+				it('should succeed', () => {
+					assert.equal(upResult2?.outcome, 'success');
+				});
+				it('should re-use stopped container', () => {
+					assert.equal(upResult2?.containerId, upResult1?.containerId);
+				});
+			});
+		});
 	});
 
 	describe('Command run-user-commands', () => {
@@ -133,6 +300,7 @@ describe('Dev Containers CLI', function () {
 	});
 
 	describe('Command exec', () => {
+
 		buildKitOptions.forEach(({ text, options }) => {
 			describe(`with valid (image) config [${text}]`, () => {
 				let containerId: string | null = null;
@@ -191,8 +359,8 @@ describe('Dev Containers CLI', function () {
 				const testFolder = `${__dirname}/configs/dockerfile-with-target`;
 				beforeEach(async () => containerId = (await devContainerUp(testFolder, options)).containerId);
 				afterEach(async () => await devContainerDown({ containerId }));
-				it('should have marker content', async () => {
-					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} cat /tmp/test-marker`);
+				it('should have build marker content', async () => {
+					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} cat /var/test-marker`);
 					const response = JSON.parse(res.stdout);
 					console.log(res.stderr);
 					assert.equal(response.outcome, 'success');
@@ -200,9 +368,9 @@ describe('Dev Containers CLI', function () {
 				});
 			});
 
-			describe(`with valid (docker-compose) config containing features [${text}]`, () => {
-				let composeProjectName: string | null = null;
-				const testFolder = `${__dirname}/configs/compose-with-features`;
+			describe(`with valid (docker-compose with image) config containing features [${text}]`, () => {
+				let composeProjectName: string | undefined = undefined;
+				const testFolder = `${__dirname}/configs/compose-image-with-features`;
 				beforeEach(async () => composeProjectName = (await devContainerUp(testFolder, options)).composeProjectName);
 				afterEach(async () => await devContainerDown({ composeProjectName }));
 				it('should have access to installed features (docker)', async () => {
@@ -221,6 +389,149 @@ describe('Dev Containers CLI', function () {
 				});
 			});
 
+			describe(`with valid (docker-compose with Dockerfile) config containing features [${text}]`, () => {
+				let composeProjectName: string | undefined = undefined;
+				const testFolder = `${__dirname}/configs/compose-Dockerfile-with-features`;
+				beforeEach(async () => composeProjectName = (await devContainerUp(testFolder, options)).composeProjectName);
+				afterEach(async () => await devContainerDown({ composeProjectName }));
+				it('should have access to installed features (docker)', async () => {
+					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} docker --version`);
+					const response = JSON.parse(res.stdout);
+					console.log(res.stderr);
+					assert.equal(response.outcome, 'success');
+					assert.match(res.stderr, /Docker version/);
+				});
+				it('should have access to installed features (hello)', async () => {
+					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} hello`);
+					const response = JSON.parse(res.stdout);
+					console.log(res.stderr);
+					assert.equal(response.outcome, 'success');
+					assert.match(res.stderr, /howdy, node/);
+				});
+			});
+
+			describe(`with valid (docker-compose with Dockerfile and target) config containing features [${text}]`, () => {
+				let composeProjectName: string | undefined = undefined;
+				const testFolder = `${__dirname}/configs/compose-Dockerfile-with-target`;
+				beforeEach(async () => composeProjectName = (await devContainerUp(testFolder, options)).composeProjectName);
+				afterEach(async () => await devContainerDown({ composeProjectName }));
+				it('should have access to installed features (docker)', async () => {
+					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} docker --version`);
+					const response = JSON.parse(res.stdout);
+					console.log(res.stderr);
+					assert.equal(response.outcome, 'success');
+					assert.match(res.stderr, /Docker version/);
+				});
+				it('should have access to installed features (hello)', async () => {
+					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} hello`);
+					const response = JSON.parse(res.stdout);
+					console.log(res.stderr);
+					assert.equal(response.outcome, 'success');
+					assert.match(res.stderr, /howdy, node/);
+				});
+				it('should have marker content', async () => {
+					const res = await shellExec(`${cli} exec --workspace-folder ${testFolder} cat /var/test-marker`);
+					const response = JSON.parse(res.stdout);
+					console.log(res.stderr);
+					assert.equal(response.outcome, 'success');
+					assert.match(res.stderr, /||test-content||/);
+				});
+			});
+
+
+			describe(`Dockerfile with post*Commands specified [${text}]`, () => {
+				let containerId: string | null = null;
+				const testFolder = `${__dirname}/configs/dockerfile-with-target`;
+				after(async () => await devContainerDown({ containerId }));
+				it('should have all command markers at appropriate times', async () => {
+					containerId = (await devContainerUp(testFolder, options)).containerId;
+					// Should have all markers (Create + Start + Attach)
+					await commandMarkerTests(testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
+
+					// Clear markers and stop
+					await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
+					await devContainerStop({ containerId });
+
+					// Restart container - should have Start + Attach
+					containerId = (await devContainerUp(testFolder, options)).containerId;
+					await commandMarkerTests(testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
+
+					// TODO - investigate what triggers postAttachCommand and check test is valid
+					// // Clear markers and re-test - should have Attach
+					// await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
+					// await commandMarkerTests(testFolder, { postCreate: false, postStart: false, postAttach: true }, 'Markers on attach');
+				});
+			});
+			describe(`docker-compose with post*Commands specified [${text}]`, () => {
+				let composeProjectName: string | undefined = undefined;
+				const testFolder = `${__dirname}/configs/compose-Dockerfile-with-target`;
+				after(async () => await devContainerDown({ composeProjectName }));
+				it('should have all command markers at appropriate times', async () => {
+					composeProjectName = (await devContainerUp(testFolder, options)).composeProjectName;
+					// Should have all markers (Create + Start + Attach)
+					await commandMarkerTests(testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
+
+					// Clear markers and stop
+					await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
+					await devContainerStop({ composeProjectName });
+
+					// Restart container - should have Start + Attach
+					composeProjectName = (await devContainerUp(testFolder, options)).composeProjectName;
+					await commandMarkerTests(testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
+
+					// TODO - investigate what triggers postAttachCommand and check test is valid
+					// // Clear markers and re-test - should have Attach
+					// await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
+					// await commandMarkerTests(testFolder, { postCreate: false, postStart: false, postAttach: true }, 'Markers on attach');
+				});
+			});
+			describe(`docker-compose with post*Commands specified (stop individual container) [${text}]`, () => {
+				let containerId: string | null = null;
+				const testFolder = `${__dirname}/configs/compose-Dockerfile-with-target`;
+				after(async () => await devContainerDown({ containerId }));
+				it('should have all command markers at appropriate times', async () => {
+					containerId = (await devContainerUp(testFolder, options)).containerId;
+					// Should have all markers (Create + Start + Attach)
+					await commandMarkerTests(testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
+
+					// Clear markers and stop
+					await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
+					await devContainerStop({ containerId });
+
+					// Restart container - should have Start + Attach
+					containerId = (await devContainerUp(testFolder, options)).containerId;
+					await commandMarkerTests(testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
+
+					// TODO - investigate what triggers postAttachCommand and check test is valid
+					// // Clear markers and re-test - should have Attach
+					// await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
+					// await commandMarkerTests(testFolder, { postCreate: false, postStart: false, postAttach: true }, 'Markers on attach');
+				});
+				describe(`docker-compose alpine with post*Commands [${text}]`, () => {
+					let containerId: string | null = null;
+					const testFolder = `${__dirname}/configs/compose-Dockerfile-alpine`;
+					after(async () => await devContainerDown({ containerId }));
+					it('should have all command markers at appropriate times', async () => {
+						containerId = (await devContainerUp(testFolder, options)).containerId;
+						// Should have all markers (Create + Start + Attach)
+						await commandMarkerTests(testFolder, { postCreate: true, postStart: true, postAttach: true }, 'Markers on first create');
+
+						// Clear markers and stop
+						await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
+						assert.ok(containerId);
+						await devContainerStop({ containerId });
+
+						// Restart container - should have Start + Attach
+						containerId = (await devContainerUp(testFolder, options)).containerId;
+						await commandMarkerTests(testFolder, { postCreate: false, postStart: true, postAttach: true }, 'Markers on restart');
+
+						// TODO - investigate what triggers postAttachCommand and check test is valid
+						// // Clear markers and re-test - should have Attach
+						// await shellExec(`${cli} exec --workspace-folder ${testFolder} /bin/sh -c "rm /tmp/*.testmarker"`);
+						// await commandMarkerTests(testFolder, { postCreate: false, postStart: false, postAttach: true }, 'Markers on attach');
+					});
+				});
+			});
 		});
 
 		describe('with valid (Dockerfile) config containing #syntax (BuildKit)', () => { // ensure existing '# syntax' lines are handled
@@ -258,6 +569,7 @@ describe('Dev Containers CLI', function () {
 			assert.equal(success, false, 'expect non-successful call');
 		});
 
+
 	});
 
 
@@ -269,11 +581,13 @@ interface ExecResult {
 	stderr: string;
 }
 
-function shellExec(command: string, options: cp.ExecOptions = {}) {
+function shellExec(command: string, options: cp.ExecOptions = {}, suppressOutput: boolean = false) {
 	return new Promise<ExecResult>((resolve, reject) => {
 		cp.exec(command, options, (error, stdout, stderr) => {
-			console.log(stdout);
-			console.error(stderr);
+			if (!suppressOutput) {
+				console.log(stdout);
+				console.error(stderr);
+			}
 			(error ? reject : resolve)({ error, stdout, stderr });
 		});
 	});