+++ CHANGELOG.md
@@ -4,6 +4,21 @@ Notable changes.
 
 ## May 2022
 
+### [0.6.0]
+
+- Handle undefined context. (https://github.com/microsoft/vscode-remote-release/issues/6815)
+- Avoid comment after ARG for Podman. (https://github.com/microsoft/vscode-remote-release/issues/6819)
+- Update to vscode-dev-containers 0.238.1.
+
+### [0.5.0]
+
+- Update to vscode-dev-containers 0.238.0.
+
+### [0.4.0]
+
+- Merge user and features Dockerfile to simplify cache and multi-platform handling.
+- Use PTY for `--log-format-json`.
+
 ### [0.3.0]
 
 - BuildKit version check for `--build-context`.